Minor fixes for Monch.

+++ src/cxx/Model/ExodusModel.cpp
@@ -1,6 +1,7 @@
 #include <Model/ExodusModel.h>
 #include <Utilities/Options.h>
 #include <Utilities/Utilities.h>
+#include <stdexcept>
 
 ExodusModel::ExodusModel(Options options) {
   mExodusFileName = options.ExodusModelFile();
@@ -47,7 +48,7 @@ void ExodusModel::initializeParallel() {
 void ExodusModel::exodusError(const int retval, std::string func_name) {
 
   if (retval) {
-    throw std::runtime_error("Error in exodus function: " + func_name + " with retval " + std::to_string(retval));
+    throw std::runtime_error("Error in exodus function: " + func_name + " with retval " + std::to_string((long long) retval));
   }
 
 }
@@ -260,7 +261,7 @@ double ExodusModel::getElementalMaterialParameterAtVertex(const Eigen::VectorXd 
   int parameter_index;
   while (!found) {
     int i = 0;
-    std::string full_name = parameter_name + "_" + std::to_string(vertex_num);
+    std::string full_name = parameter_name + "_" + std::to_string((long long) vertex_num);
     for (auto &name: mElementalVariableNames) {
       if (name == full_name) {
         parameter_index = i; found = true;
